+++ contracts/p0/Furnace.sol
@@ -13,32 +13,32 @@ contract FurnaceP0 is ComponentP0, IFurnace {
     using FixLib for uint192;
 
     uint192 public constant MAX_RATIO = 1e18;
-    uint32 public constant MAX_PERIOD = 31536000; // {s} 1 year
+    uint48 public constant MAX_PERIOD = 31536000; // {s} 1 year
 
     uint192 public ratio; // {1} What fraction of balance to melt each period
-    uint32 public period; // {seconds} How often to melt
-    uint32 public lastPayout; // {seconds} The last time we did a payout
+    uint48 public period; // {seconds} How often to melt
+    uint48 public lastPayout; // {seconds} The last time we did a payout
     uint256 public lastPayoutBal; // {qRTok} The balance of RToken at the last payout
 
     function init(
         IMain main_,
-        uint32 period_,
+        uint48 period_,
         uint192 ratio_
     ) public initializer {
         __Component_init(main_);
         setPeriod(period_);
         setRatio(ratio_);
-        lastPayout = uint32(block.timestamp);
+        lastPayout = uint48(block.timestamp);
         lastPayoutBal = main_.rToken().balanceOf(address(this));
     }
 
     /// Performs any melting that has vested since last call.
     /// @custom:refresher
     function melt() external notPausedOrFrozen {
-        if (uint32(block.timestamp) < uint64(lastPayout) + period) return;
+        if (uint48(block.timestamp) < uint64(lastPayout) + period) return;
 
         // # of whole periods that have passed since lastPayout
-        uint32 numPeriods = (uint32(block.timestamp) - lastPayout) / period;
+        uint48 numPeriods = (uint48(block.timestamp) - lastPayout) / period;
 
         // Paying out the ratio r, N times, equals paying out the ratio (1 - (1-r)^N) 1 time.
         uint192 payoutRatio = FIX_ONE.minus(FIX_ONE.minus(ratio).powu(numPeriods));
@@ -53,7 +53,7 @@ contract FurnaceP0 is ComponentP0, IFurnace {
 
     /// Period setting
     /// @custom:governance
-    function setPeriod(uint32 period_) public governance {
+    function setPeriod(uint48 period_) public governance {
         require(period_ > 0 && period_ <= MAX_PERIOD, "invalid period");
         emit PeriodSet(period, period_);
         period = period_;
